Crate pliantdb_counter_client

Structs

Counter
The state of the Counter component.

DatabaseContext
A context provides the information necessary to communicate with the user inteface.

Enums

CounterEvent
All events that the Counter behavior will receive from child widgets.

CounterWidgets
This enum identifies widgets that you want to send commands to. If a widget doesn’t need to receive commands, it doesn’t need an entry in this enum.

DatabaseCommand
Commands that the user interface will send to the database task.

Functions

increment_counter

main
The example’s main entrypoint.

process_database_commands
Processes each command from receiver as it becomes available.

watch_for_changes
Listens for PubSub events that come in from other clients pressing the button.